requirements.txt
streamlit
opencv-python-headless
tensorflow-cpu
numpy
streamlit-webrtc
av